Platform consistency fixes, performance and misc. improvements.
File.Delete does not throw anymore if path points to dir since it does not throw if the file does not exist.
Added optional overwrite parameter to IAbsoluteFilePath.MoveTo() that can be used on .NET Core 3+.
Doc fixes to clarify which path format is used when combining string path with dirs.